Ivi_test_project/requirements.txt
requests
pytest
pytest-xdist
allure-pytest
dataclasses_json
Faker
python-dotenv